_threadmodule: thread-safety fixes
 - Make rlock thread-safe
 - Use atomics to modify interp->num_threads

+++ Include/internal/pycore_interp.h
@@ -92,7 +92,7 @@ struct _is {
         /* The linked list of threads, newest first. */
         PyThreadState *head;
         /* Used in Modules/_threadmodule.c. */
-        long count;
+        Py_ssize_t count;
         /* Support for runtime thread stack size tuning.
            A value of 0 means using the platform's default stack size
            or the size specified by the THREAD_STACK_SIZE macro. */
